Remove 'error' event & add 'stream-loaded' event
When the camera stream is not loaded yet, the canvas has zero width.
This in turn throws exceptions in scan() due to getImageData calls. To
avoid this, scanning now starts when the stream is actually loaded. This
is also passed as an event.

The error event on the other hand is removed. It feels weird to handle
unexpected errors this way.