Change `locate` event payload scheme
Instead of passing an object, representing the position of QR code
modules, an array of detected positions is propagated. If nothing
was detected, the payload contains an empty array. This removes the
need for `null` checks on listener side. Positions can just be iterated
and painted on the way.

Also improved browser support for assignment of camera streams to video
elements by adding fallback cases.