Join several events in new 'init' event
Both the 'no-support' and 'permission-deny' event mark fatal errors
preventing the 'stream-loaded' event from emitting. Displaying a loading
indicator based on those events is messy. With the unified 'init' event
this is easier. It's also more flexible and cleaner to implement. Only
one event and therefore only one handler has to be registered.