Fix error on component destroy
When the component is destroyed, the scan loop is stopped. Because each
iteration executes an asynchronous function (via 'requestAnimationFrame')
the final iteration has to deal with an component instance which is
already partly destroyed. 'captureFrame' is called in each iteration and
accesses the video and canvas ref. Those are not available anymore in the
final scan step causing an error. To avoid this 'captureFrame' is now
called outside the 'requestAnimationFrame' callback.