refactor: rename QrcodeReader to QrcodeStream
Since the project now exposes multiple components, it's worth
distinguishing between the name of the project and the name of the
components. QrcodeReader now refers to the project, while
QrcodeStream refers to the component which detects and decodes
qrcodes from a camera stream.

In contrast to QrcodeReader, QrcodeStream does not expose the
deprecated `video-constraints` prop anymore and also does not
detect drag-and-dropped images/urls anymore (this feature can be
easily re-added by wrapping QrcodeStream with QrcodeDropZone).