fix(QrcodeStream): stop worker on destroy
The worker which is processing scanning jobs, is not terminated
when the component is destroyed. This was intended to be
guaranteed by the following mechanism:

1. Component is destroyed (for example with `v-if="false"`)
2. `beforeDestroy` hook is called
3. `this.destroyed = true`
4. computed property `shouldStream` becomes `false`
5. computed property `shouldScan` becomes `false`
6. `shouldScan` watcher calls `stopScanning`

Step 4 is never reached though. Probably becomes Vue won't
re-evaluate computed properties after the `beforeDestroy` hook
was called. So `stopScanning` is never called and the worker keeps
processing frames. Results are still handed to the component which
tries to paint to a canvas which doesn't exit anymore (since the
component is destroyed). Hence causing errors like

_Cannot read property 'getContext' of undefined_

See #85