perf: ignore color inverted QR codes
Standard QR codes have black modules on white background. If an
image can't be decoded it has to be re-processed to check if it
might be a QR code with light modules on dark background. Checking
for this edge case is not worth the performance loss though.

BREAKING CHANGE: Color inverted QR codes are not detected anymore.