fix(tracking): don't translate width/height
Camera resolution does not match the dimensions of the rendered
component. So when a QR code is detected we need to adjust it's
coordinates before it's passed to the track function. For that
xy-coordinates need to be scaled to account for responsiveness and
they need to be translated to account for cropping (especially in
fullscreen mode).

However, widths/heights don't need to be translated since they only have
magnitude but no position. That error caused tracking to messed up in
fullscreen mode for tracking functions that use width/height in the
boundingBox property.

Issue: #244