feat: merge master into vue3-compatibility (#314)
* fix(QrcodeStream): iOS 15 won't render camera

When a camera stream is loaded, we assign the stream to a `video`
element via `video.srcObject`. At this point the element is hidden with
`v-show="false"` aka. `display: none`. We do that because at this point
the videos dimensions are not known yet. We have to wait for the
`loadeddata` event first. Only after that event we display the video
element. Otherwise the elements size awkwardly flickers.

However, it appears since iOS 15 all iOS browsers won't properly render
the video element if the `video.srcObject` was assigned *while* the
element was hidden with `display: none`. Using `visibility: hidden`
instead seems to have fixed the problem though.

Issue: #264 #266

* fix(QrcodeStream): black list cameras by label

Modern devices sometimes have multiple rear cameras. Not all are optimal
for scanning QR codes. For example wide angle, infra-red and virtual
cameras. By maintaining a black list of unsuitable cameras with unique
labels we can gradually exlude more and more of them.

Issue: #269 #253

* fix(QrcodeStream): reject infrared cameras

When automatically selecting a camera we want to avoid infrared cameras
since they are not suitable for scanning QR codes. Some infrared camera
labels have the substring "infrared". So we filter by that.

Issue: #269

* docs: update Upload demo

On supporting mobile devices QrcodeCapture does not open the file dialog
by default but starts the camera instead. This is supposed to be a
feature but developers keep being surprised by it or think it's a bug.
Therefore trying to clarify this in the associated demo.

Issue: #167 #211 #232 #272 #278

* feat(ts): specify types in package

* Remove unused obsolete import

DropImageDecodeError was removed in https://github.com/gruhn/vue-qrcode-reader/commit/dea620ec69f0644c3c7282aff08686c8f639edbb

Co-authored-by: Niklas Gruhn <niklas@gruhn.me>
Co-authored-by: RobWalker <rob.walker@auctria.com>