fix(QrcodeStream): ignore structually equal prop changes
Props like `constraints` and `formats` which carry non-primitive values
might receive structurally equal updates. For example, let `constraints`
be the variable that is passed to `QrcodeStream`:

    <qrcode-stream :constraints="constraints" />

and imagine the `script` section looks like this:

   const constraints = ref({})

   setInterval(() => {
     constraints.value = { deviceId: 'whatever' }
   }, 100)

This would keep triggering updates in `QrcodeStream` although the
constraints don't actually change. This is because the assigned object
is referencially different every time and Vue only checks referencial
equality. A less contrived example where this happens, is when the
template looks like this:

    <qrcode-stream :constraints="{ deviceId: 'whatever' }" />

Whenever Vue re-evaluates the passed object it creates a referencially
different copy. To avoid this problem we now maintain "cached" versions
of these props and only update them when we detect strucural changes.