fix: native `BarocdeDetector` doesn't support requested formats
Previously we always used the native `BarocdeDetector` implementation if
available. However, the native API might be available but does not
support the barcode formats requested by the user. For example, `"qr_code"`
might be supported but the user wants to scan `["qr_code", "aztec"]`.
In these cases, we now also fallback to the polyfill implementation.

Closes #450